requirements.txt
fastapi==0.103.0
sqlalchemy==2.0.34
sse_starlette==2.1.3
azure-storage-blob==12.22.0
psycopg2-binary==2.9.9
pypdf == 5.0.0
pytz==2024.1
requests==2.32.3
pandas==1.5.3
numpy==1.24.3
fuzzywuzzy==0.18.0
opencensus==0.11.4
opencensus-ext-azure==1.1.13
Authlib==1.3.2
email-validator==2.2.0
passlib==1.7.4
python-Levenshtein==0.21.1
reportlab==4.2.2
pdf2image==1.17.0
python-multipart==0.0.9
openai==1.42.0
Office365-REST-Python-Client==2.3.13
opencv-python-headless==4.10.0.84
python-magic-bin==0.4.14
scikit-learn==1.5.2
azure-keyvault-secrets==4.8.0
azure-ai-formrecognizer==3.3.3
pydantic_settings==2.5.2
alembic==1.13.2
pre-commit==3.8.0
detect-secrets==1.5.0
types-pytz==2024.2.0.20240913
types-requests==2.32.0.20240914
